Fix the Vishay Minicast footprint
The radius of the receiver part is 2 mm, not 1 mm.  I can't find a
specific datasheet for the Minicast footprint, but its dimensions can be
seen e.g. in https://www.vishay.com/docs/82491/tsop382.pdf

A few other changes were made as well to bring the footprint up to
current KLC.